pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-llm-social-agent"
version = "0.1.0"
description = "Telegram-bot-driven llm-social-agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-telegram-bot>=21.0",
  "python-dotenv>=1.0.0",
  "pyyaml>=6.0",
  "requests>=2.31.0",
  "requests-oauthlib>=1.3.1",
  "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
